This has following fixes. Apart from these default user name is changed from praba to root 1.For large allocator (size >PAGE_SIZE) first page is always empty [1844630 ] This is because it does not check the curPage for free space. It directly allocaes new page from database.
2.Chunk created for storing global hash index nodes is
never used. [1844632 ]
This chunk creation is removed as every index will
have its own chunk to allocate from. This is done so
that high concurrency is attained.

3.catalog -D trans does not display total undo log
nodes and lock nodes [1844634 ]

4.insertnulltest1.c failing [1844638 ]
This is because of wrong usage of >operator instead of <operator, when deciding
whether to store nullbits in integer or char array

5.when sysdb gets full because of huge undo logs created for large transactions, it dumps core
   undo allocator does not check the return value of chunk->allocate [ 1761385 ]

6.when field is not binded and notnull bit is not set, it is not inserting null. [1844642 ]
Marking the field null bit during TableImpl::copyValuesFromBindBuffer if it is not binded
7.NOT NULL constriant support for storage engine 1692204