fix for Bugs: 1955997, 1955999, 1956001/03/11/20/21/23/24.
Memory leaks are plugged in reset function.

InsertPredicate function is changed where Condition object was created,
which is a wrapper for predicateImpl.  Directly PredicateImpl is used.
Condition object was dumping the core as condition goes out of scope in
this function, causing the PredicateImpl object to go out of scope.