Implement Predicate (un)serialization.
Most notable change is the introduction of PredicateImpl::Serialized as
an intern serialization structure. We save quite a bit space when
serializing this subset of PredicateImpl.

Also important is the change to have CGRANT store a raw pointer to the
Page data where the serialized Predicates/ConditionValues end up in.

There are still some TODOs like taking locks and most importantly, not
allocating a Page for every CGRANT structure. That's pure overkill.